tratamento de exceções

+++ inspecao-service/inspecao-service/src/main/java/com/infnet/inspecao_service/service/InspecaoExtintorService.java
@@ -15,27 +15,37 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class InspecaoExtintorService {
 
-
     private final InspecaoExtintorRepository inspecaoExtintorRepository;
     private final RestTemplate restTemplate; // Comunicação com Extintor-service
 
     public List<InspecaoExtintor> findAllInspecoes() {
         log.info("Buscando todas as inspeções");
-        return inspecaoExtintorRepository.findAll();
+        List<InspecaoExtintor> inspecoes = inspecaoExtintorRepository.findAll();
+        if (inspecoes.isEmpty()) {
+            throw new IllegalArgumentException("NÃO EXISTE NENHUMA INSPEÇÃO CADASTRADA!");
+        }
+        return inspecoes;
     }
 
     public Optional<InspecaoExtintor> findInspecaoById(String id) {
         log.info("Buscando inspeção pelo ID: {}", id);
-        return inspecaoExtintorRepository.findById(id);
+        Optional<InspecaoExtintor> inspecao = inspecaoExtintorRepository.findById(id);
+        if (inspecao.isEmpty()) {
+            throw new IllegalArgumentException("INSPEÇÃO NÃO ENCONTRADA!");
+        }
+        return inspecao;
     }
 
     public List<InspecaoExtintor> findInspecoesByExtintorId(String extintorId) {
         log.info("Buscando inspeções pelo ID do extintor: {}", extintorId);
-        return inspecaoExtintorRepository.findByExtintorId(extintorId);
+        List<InspecaoExtintor> inspecoes = inspecaoExtintorRepository.findByExtintorId(extintorId);
+        if (inspecoes.isEmpty()) {
+            throw new IllegalArgumentException("NENHUMA INSPEÇÃO ENCONTRADA PARA O EXTINTOR!");
+        }
+        return inspecoes;
     }
 
     public InspecaoExtintor saveInspecao(InspecaoExtintor inspecaoExtintor) {
-
         // Validação: Verificar se o extintor está cadastrado no Extintor-service
         String url = "http://extintor-service/extintores/" + inspecaoExtintor.getExtintorId();
         Boolean extintorExiste = restTemplate.getForObject(url, Boolean.class);
@@ -54,7 +64,6 @@ public class InspecaoExtintorService {
             inspecaoExtintor.setStatus("Não conforme");
         }
 
-
         return inspecaoExtintorRepository.save(inspecaoExtintor);
     }
 
@@ -66,13 +75,16 @@ public class InspecaoExtintorService {
                 && inspecao.isSuporteBoasCondicoes() && inspecao.isLacreIntacto();
 
         //LocalDate hoje = LocalDate.now();
-       // boolean validadeExtintorValida = hoje.isBefore(inspecao.getDataInspecao());
+        // boolean validadeExtintorValida = hoje.isBefore(inspecao.getDataInspecao());
 
         return camposValidos /*&& validadeExtintorValida*/;
     }
 
     public void deleteInspecao(String id) {
         log.info("Deletando inspeção com ID: {}", id);
+        if (!inspecaoExtintorRepository.existsById(id)) {
+            throw new IllegalArgumentException("INSPEÇÃO NÃO ENCONTRADA!");
+        }
         inspecaoExtintorRepository.deleteById(id);
     }
 }
